# openapi.yaml
openapi: 3.1.0
info:
  title: Pet Store API
  description: API for managing pets in a pet store
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local development server
paths:
...
  /cart:
    get:
      summary: Get items in the current user's cart
      operationId: getCart
      security:
        - googleOAuth: [ user ]
      parameters:
        - name: Authorization
          in: header
          required: true
          description: Bearer token obtained from Google OAuth (e.g., "Bearer eyJhbGci...")
          schema:
            type: string
      responses:
        '200':
          description: List of items in the cart
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CartItemView'

    post:
      summary: Add a pet to the current user's cart
      operationId: addToCart
      security:
        - googleOAuth: [ user ]
      parameters:
        - name: Authorization
          in: header
          required: true
          description: Bearer token obtained from Google OAuth (e.g., "Bearer eyJhbGci...")
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CartItemPost'
      responses:
        '200':
          description: Pet added to cart
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartItem'
...
components:
  schemas:
...
    CartItem:
      type: object
      properties:
        id:
          type: string
          description: ID of the cart item
          x-displayName: ID
        userId:
          type: string
          description: ID of the user who owns this cart
          x-displayName: User ID
        petId:
          type: string
          description: ID of the pet added to the cart
          x-displayName: Pet ID
    CartItemPost:
      type: object
      properties:
        petId:
          type: string
          description: ID of the pet added to the cart
          x-displayName: Pet ID
    CartItemView:
      type: object
      properties:
        id:
          type: string
          description: Cart item ID
          x-displayName: ID
        petName:
          type: string
          description: Name of the pet in the cart
          x-displayName: Name
        quantity:
          type: integer
          description: Number of times this pet is in the user's cart
          x-displayName: Quantity
...
  securitySchemes:
    googleOAuth:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://accounts.google.com/o/oauth2/auth
          tokenUrl: https://oauth2.googleapis.com/token
          scopes:
            guest: "Access as a guest (no authentication required)"
            user: "Registered user access"
            admin: "Administrator access"